BUG#22350047: IF CLIENT KILLED AFTER ROLLBACK TO SAVEPOINT               PREVIOUS STMTS COMMITTED
ANALYSIS:
=========
Consider an uncommitted transaction in which we have
created a savepoint and executed
'ROLLBACK TO SAVEPOINT <savepoint_name>' and killed the
client. When another client is started, the changes that
were made before the savepoint creation are still
reflected implying that the transaction was committed.

In this scenario, when the client is killed, THD::cleanup()
is called. During the cleanup, since the binary log is
enabled, MYSQL_BIN_LOG::rollback() is called. The logic
inside this function is, If thd->lex->sql_command ==
SQLCOM_ROLLBACK TO SAVEPOINT, we should not perform a
engine rollback, instead only the caches need to cleared.
Here since the last executed command before the client was
killed was 'ROLLBACK TO SAVEPOINT', thd->lex->sql_command
will contain SQLCOM_ROLLBACK_TO_SAVEPOINT and hence engine
rollback is not done and later the transaction is
committed. Thus the changes done before the savepoint
creation is reflected.

This issue was caused because resetting of
thd->lex->sql_command was not done after the execution
of each statement.

FIX:
====
We are now setting thd->lex->sql_command to SQLCOM_END
after execution of each statement.