BUG#25673788: MYSQLADMIN SHUTDOWN HANGS WHEN GR PLUGIN IS BUSY
When we try to shutdown server through mysqladmin when GR plugin is in
deleted state, shutdown hangs.

When the plugin is in use and the server tries to uninstall it, it will mark
the plugin as deleted and delay the uninstall to when execute shutdown.
When executing the shutdown the server don't execute anything on plugins
deleted.

To solve this we don't allow remove the plugin when this is in use, it need to
be executed first the command STOP GROUP_REPLICATION.