BUG#25896166 APPLYING_TRANSACTION.R..A..S_BY_WORKER COL DOESN'T GET CLEARED FOR GR RECOVERY
Problem:
Sporadically, when MTS is enabled, the value for APPLYING_TRANSACTION in
the P_S table replication_applier_status_by_worker would not be empty
for an inactive worker.

Analysis:
The structure which populates the table corresponding to the worker's
thread after it is no longer running was being populated before the
thread stopped. This means that, sporadically, the information on the
table did not display the correct state of the thread.

Fix:
Wait for the worker thread to stop before copying the information to the
structure that is used to populate the P_S table.