BUG#26277771 - BAD WRITE SET TRACKING WITH UNIQUE KEY ON A DELETE FOLLOWED BY AN INSERT
Problem: When generating writesets the collations on the database were
not properly checked, so when case/accent insensitive collations are
used on the database the writesets would be different for keys the
database considers to be the same.

Solution: The fix for this issue consists of generating the writesets
from the keys only after addressing any collation effects using strnxfrm.