BUG#25982097 WRITESET DEPENDENCY TRACKING SUB-OPTIMAL AFTER ROTATION OR              ON EMPTY TRANSACTIONS
This is a post push fix to easy the backporting.

The previous approach of detecting an empty transaction was counting the
amount of events in the binlog cache.

We changed this approach to consider two new flags: one if the cache has
seen a transaction start statement and other if the cache has seen a
transaction end event. A transaction cache that saw a start, an end and
didn't has statements or rows changing content will be considered having
an empty transaction.