BUG#26571683: test gr_applier_early_failure is failing
The test was failing, not being able to DROP table used on test.

The cause was START GROUP_REPLICATION can return success and failure, so we
skip the reset of super read only when execute STOP GROUP_REPLICATION to
prevent the case that START fail.

We disable the super read only after STOP to prevent failure of dropping table
when START succeeds.