Bug#26368004: CONFUSING GR ERROR MESSAGE: CHECK CREDENTIALS
Description:
If hostname resolution does not work for a member in Group Replication, the
error returned when attempting to connect is:

2017-06-29T00:56:02.103512Z 23 [ERROR] Plugin group_replication reported:
'There was an error when connecting to the donor server. Check group
replication recovery's connection credentials.'

But the issue has nothing to do with credentials, but the fact that Group
Replication only works when the hostname/IP address is resolvable.

Resolution:
Error message has been improved.