Bug#26519582: GR_SINGLE_PRIMARY_ASYNC_AFTER_GR_SECONDARY FAILING ON HUDSON
Issue:
======
The testcase failing because of concurrent DDL statement execution. The change
master which is executed on server2 is blocked by create table (initially
executed on server1) when applied on server2 as it had higher priority.

Solution:
=========
Move create table on server1 below change master statement on server2, so they
are not executed concurrently.