Bug#26614455: MAKE CACHED JSON_PATH OBJECTS IMMUTABLE
The cached Json_path object returned by Json_path_cache::get_path() is
mutable. If a caller modifies the returned object and doesn't restore
the original state of the object when it is done, the cache is
corrupted and could cause wrong results the next time it is used.

JSON_SEARCH is the only caller that does any modification on the
cached Json_path object. This patch changes it so that it doesn't
modify the Json_path object. Instead it makes the modifications on a
String object that represents the path. Since the path needs to be
converted to a string in the end anyway, this saves some round-trips
between Json_path representation and String, and gives a small speedup
as an extra bonus.

Json_path_cache::get_path() is changed to return a pointer to a const
Json_path object.

Additional changes in the JSON_SEARCH function, while at it:

Remove the String members m_one_or_all_value and m_escape. Since the
valid values of the one-or-all argument and escape character argument
of JSON_SEARCH are quite small (3 characters and 1 character), there
isn't any point in caching these strings in the Item object. Creating
a small buffer on the stack would be enough to avoid heap allocation
in all the common cases.

Microbenchmarks (64-bit, Intel Core i7-4770 3.4 GHz, GCC 6.3):

BM_JsonSearch              1389 -> 1163 ns/iter [+19.4%]
BM_JsonSearch_Wildcard     3203 -> 3216 ns/iter [ -0.4%]

Change-Id: I31c5adaafc41403efcbc1edb84ec76b2b7949b57