Bug#26272084 DYNAMIC LINKING WITH OPENSSL FOR GENERIC LINUX
Add support for linking with libssl.so and libcrypto.so on Linux
This will copy libssl.so and libcrypto.so into library_output_directory in the
build tree, and use them when building our executables and shared libraries.
The openssl .so libraries will also be INSTALLed together with our own .so
libraries. This is done iff:

- cmake is run with -DWITH_SSL=/path/to/installation
- /path/to/installation contains the necessary .so libraries
- this is a LINUX build
- this is a STANDALONE INSTALL_LAYOUT