Bug#26497353: ASSERTION FAILED: M_OPENED_TABLE != NULLPTR
Problem:
Temp table is not created to handle grouping for the query as its grouping
on constants. When a temp table is created for handling windowing function,
it does not get instantiated because window_short_circuit is enabled and we
choose end_write_group to write to it. But end_write_group fails while trying
to write to the uninstantiated temp table.

Analysis:
For the given query, a temp table is needed to materialize group rows for rollup
before windowing functions can be handled. But need_tmp_before_win does not
get set. If rollup is not present, "group by" gets optimized away in
optimize_distinct_group_order() and we do not see the issue. However, when
rollup is present, as group by is not optimized away for this case, we need a
temp table. (see optimize_distinct_group_order())

Solution:
Set need_tmp_table_before_win, when rollup is present with windowing functions
(similar to rollup with distinct).