Bug #26496645: WINDOW FUNCTIONS: CRASH IN WINDOW::RESTORE_SPECIAL_RECORD
Problem
When we have a value which is less than argument value for a range frame in
a windowing function, mysql server exits

Analysis:
Currently, for a range optimizable window function, the result of the
window function is stored into a in-memory table in the form of a
special record, which can be restored later.
This saves us time from calculating the results again for the next row if
the value for window function is expected to be the same. But for a case
when all the values in the frame are less than the specified range value,
we skip saving this special record. This results in restoring a non-existent
record and the server exits because of this.

Solution:
Added saving of the special record to the code flow in
 process_buffered_windowing_record.