Bug #26026218 : TRANSPORTER ERROR 0X8004, 0X8023; CHECKSUM; UNSUPPORTED BYTE ORDER
Signals are integer numbers of 32-bit words long.
A message checksum is the bitwise XOR of each of these words, so that the checksum
word XORed with the other words in the signal gives zero.

Current checksum utilities operate only on integer numbers of 32 bit words, but
it is useful to be able to build a checksum cumulatively using arbitrary spans
of bytes, which may not have corresponding relative word alignment.

This is done by :
 - Observing that the word XOR is a series of byte XORs, from byte 0 to byte 3
 - Observing that the 'next byte to XOR' can be memorised in the checksum
   itself by rotating the checksum so that byte 0 is always next.
   - When a multiple-of-4 bytes have been processed, the checksum has rotated back
     to its original alignment.

Steps :
 1.  Determine word-alignment of the source bytes
 2.  XOR in any unaligned partial-word bytes at the start
 3.  Rotate checksum to be used for word-aligned XOR of the bulk of the byte
     stream
 4.    Use normal word-oriented checksum for the bulk
 5.  Rotate checksum back to start position
 6.  XOR in any unaligned partial-word bytes at the end
 7.  Rotate checksum to be correct w.r.t the number of bytes added in