Bug #25892493 GR XCOM SERVICE SEGFAULTS WHEN MORE THAN 1024 FILE DESCRIPTORS ARE IN USE
Switched from select to poll in task.c and inlined functions in
xdr_utils. The polling code now dynamically allocates as space for new
file descriptors, so the number is now in principle only limited by
the available memory. Added deinit functions. Fixed bugs in unpoll and
wake_all_io that caused task waiting for IO to not terminate because
the reference count never went to 0.
Added interface to WSAPoll on Windows.