Bug#26497247:WINDOW FUNCTIONS: CRASH IN DO_COPY_MAYBE_NULL
Problem:
Writing NULL to a non-nullable column results in server exit as null_ptr is not
allocated.

Analysis:
ROLLUP writes NULL's to columns which are part of group by clause. So we set
"maybe_null" to true in ::resolve_rollup() if any item in the item
list is part of group_by clause. We also set "maybe_null" to true for all the
arguments to a function, which are part of group_by clause. This happens in
::change_group_ref() and similarly we generate Item_null_result objects
in ::rollup_make_fields() for items which are part of group by clause.

However, optimizer does not set "maybe_null" to true for arguments to a
window function as ::change_group_ref() is called only if an item is of
Item_func type which is not the case for windowing functions.
But we do add all the arguments of a window function to the item list when
::split_sum_func() gets called. So checking if this argument is equal to any
item in group by clause, would ensure that "maybe_null" is set to true.
By setting it to true, fields would be created in temp tables using
create_tmp_field_from_item() instead of create_tmp_field_from_field() which
further ensures that null_ptr's are allocated correctly.

Solution:
Check if an item is part of group by clause by calling ::eq() instead
of just pointer comparison. This is in line with what we do in ::change_group_ref()
and in ::rollup_make_fields().