Bug #26026218 : TRANSPORTER ERROR 0X8004, 0X8023; CHECKSUM; UNSUPPORTED BYTE ORDER
A checksum checking class is added to the Transporter, allowing checksums to
be verified between the application point-of-send (sendSignal et al) and the
point of passing data to the OS.

This involves parsing the sequences of bytes made available to send, to find
message boundaries, and checksumming and verifying the content.

In case of failure, as much context as is available at the point of failure
is dumped out.