BUG#26241291 EXCESSIVE MEMORY USAGE BY XCOM WITH LARGE INSERT WORKLOADS
Problem
-------
In MySQL 5.7.x, XCOM controls the cache's size by defining a maximum number of
entries and only replaces old entries to accommodate new ones. It never purges
unnecessary and old entries if there is no need to make room for new ones. This
behavior along with the fact that the size of each entry is never taking into
account lead to excessive memory consumption.

This behavior can also make nodes slow thus leading them to be expelled and
eventually causing a majority loss which blocks the whole system.

Fix
---
The fix consists of limiting the memory used by the nodes and automatically
purging old and unused entries. Before doing so, the nodes must agree on the
messages that they must keep in the cache to avoid purging messages that are
not received yet by all nodes. Periodically, nodes send the sequence of the
last message they got an agreement on and define which messages can be
discarded.

This additional information exchanged by them implies a change to the XCOM
protocol version. Because of this, a bug in the protocol version negotiation
had to be fixed as well.

The fixes were already pushed to trunk and were backported 5.7:

BUG#22671683: LIMIT MEMORY UTILIZATION IN XCOM

BUG#25568493: GROUP MEMBERS BECOMES UNREACHABLE WHEN SERVER WITH LOWER
              VERSION TRIES TO JOIN