Bug #26026218 : TRANSPORTER ERROR 0X8004, 0X8023; CHECKSUM; UNSUPPORTED BYTE ORDER
Checksum checking is added to the TCP Transporter using the new class + utils.

doSend() obtains an array of IOVECs and attempts to send all of the data from
them, in multiple passes.

The OS may manage to send some, all or none of the data, which could result in
IOVEC offsets changing, becoming odd etc over time.

doSend() will :
  a) Check checksums for all data available in IOVECs prior to OS::send()
  b) Check checksums for all sent data afer OS::send()

Obviously this adds some CPU load to the threads performing sending.