Post-push fix: WL#2955: RBR replication of partial JSON updates
Fix sporadic test failure.

Variable $fix in rpl_row_jsondiff_stress may be empty; when that
happens the line '--eval $fix' failed.  It is expected and allowed
that $fix is empty sometimes, just a bit unlikely due to the ranomized
nature of this test, and therefore the mistake was not found earlier.
Fixed by guarding '--eval $fix' with 'if ($fix)'.