Bug#25817660: Combination of virtual index, foreign key and trigger asserts
The problem is with handler::my_eval_gcolumn_expr_with_open().
This function is designed to work from purge threads, but here it is
used in regular execution of foreign key processing.
From a purge thread it needed to initialize the current LEX before
evaluating a generated column. But in regular processing, this will
destroy the current statement context, causing an assert error
in dispatch_command().

The fix is to remove the calls to lex_start() and lex_end() in
handler::my_eval_gcolumn_expr_with_open(). This is safe since neither
of thd->dd_client()->acquire(), open_table_uncached() nor
my_eval_gcolumn_expr_helper() affect the current LEX object.

Apparently, this is also safe when called from a purge thread.
Existing tests are unaffected by this change.