Bug#26649978: JSON_INSERT: ASSERTION FAILED: PARENT
An assertion was raised if JSON_INSERT was used to insert an element
into the root array or object of a JSON document, and auto-wrapping
was used to identify the root element using the given path.

The code that checks if we are inserting into the root element, checks
if the length of the path is one. That is, it checks that the path
consists of a single path leg that tells which position the inserted
element will have inside the root element. But if there are
auto-wrapping path legs at the beginning of the path, a path whose
length is greater than one might also refer to an element in the root,
so checking the path length doesn't reliably tell us if we're
inserting into the root.

This was not a problem until bug#26022576 made JSON_INSERT aware of
auto-wrapping.

The fix is to change the detection of the root element to check if the
matched element has a parent. If it has no parent, it is the root
element.

For consistency, also make JSON_SET check if the element has a parent
instead of checking the length of the path. In this case using the
length of the path does not lead to any problems, since JSON_SET
removes redundant auto-wrapping path legs before the path is used.

Change-Id: Icb719031662700f213d3b31c5fa144aa7b8310b7