Bug#25510805 - MYSQL CRASHES WHEN TRYING TO CONNECT FROM                A HOST WITH AHOSTNAME OF 69 CHARACTERS
DESCRIPTION:
===========
When a connection is made from a host with a hostname
length of more than HOSTNAME_LENGTH characters,
MySQL crashes with a buffer overflow.

ANALYSIS:
========
If the hostname of the connecting host is greater than
HOSTNAME_LENGTH (60 chars), copying the hostname to the
performance schema table field which is limited to
HOSTNAME_LENGTH characters results in a crash.

FIX:
===
The fix is to truncate the hostname to HOSTNAME_LENGTH
number of characters before adding them to the performance
schema tables which record host information and details.
This makes sure that the client connection is established
successfully.