Bug #26696147: REMOVE MY_SNPRINTF
Remove my_snprintf, replacing it with C++11 snprintf everywhere. Fix a myriad
of compiler warnings uncovered as a result, several of them real bugs.

The old (non-standard) behavior of %s was to stop at the first invalid UTF-8
codepoint.  While useful in specific situations, it was also confusing in
others (using e.g. a binary string as INTEGER would print out an empty string,
instead of showing that there was data in the string), and thus has been
replaced with showing question marks for invalid UTF-8.

Change-Id: Ica316ff9d32053c6dc018037cd0d1645ab319d45