Bug#25475132 GROUP REPLICATION: MYSQL CLIENT CONNECTIONS HANG DURING GROUP REPLICATION START Bug#26037344 THE DELAYED PLUGIN INITIALIZATION THREAD IS LEFT RUNNING WHEN IT IS NOT NEEDED
Client complaints existed about how group replication would make
connections hang on server start if the plugin was also
starting. i.e., joining the group.
To solve this we reworked the delayed initialization mechanism used on
server starts, so it only blocks connections until the server is in
read mode, not waiting for the member to join the group as before.
This rework also solves the issue of the thread responsible for this
task being initialized and running even when not needed.