BUG#26611272: Test gr_uninstall_while_network_blocked  is failling on PB2
Desciption:

This test create a group of two members, where one member is put in a UNREACHABLE state
and then INSERT is executed which gets blocked, then when we try to execute UNINSTALL
command it returns an error that plugin cannot be uninstalled now.

But if the UNINSTALL is executed before the plugin receives INSERT query, the plugin
will be able to UNINSTALL successfully, that results in a test failure as is
expecting UNINSTALL to fail with error ER_PLUGIN_CANNOT_BE_UNINSTALLED.

Solution:

To prevent the uninstall getting executed successfully, we wait for server to receive
the INSERT, before executing the UNINSTALL of Group Replication plugin.