Bug#26377159: IMPLEMENT DOWN_CAST CORRECTLY
The function template down_cast in its current form only
works for pointers, but the template as written will match
any type. In order for it to work for references (lvalue or
rvalue), a different implementation is needed.

Fixed by limiting the current down_cast to pointers and
adding an overload for references. This also has the added
benefit of not matching for non-pointer non-reference types.