Bug#26024253: GR MEMBER STOP MUST STOP ALL ASYNCHRONOUS CHANNELS THAT ARE FEEDING DATA INTO IT
Issue
=====
If a member that has asynchronous channels feeding data into it - either on
multi-primary deploy or the primary on single-primary deploy - does stop, all
asynchronous channels must be stopped to avoid that changes are made locally
(not replicated to the group).

Even when super_read_only it is set on STOP GROUP_REPLICATION, that will not
affect already running asynchronous which can still do changes.

Solution
========
When group replication stops, due to error or due to executing stop
group_replication command, all asynchronous channels are stopped.