Bug #26665851 - USING FLUSH TABLES WITH READ LOCK AND LOCK INSTANCE LEADS TO A CRASH
The problem was that FLUSH TABLES WITH READ LOCK statement could trigger
a debug  assert if the current connection had acquired BACKUP LOCK.
This had no consequences for release builds.

This patch fixes the problem by adjusting the assert so
that it takes the possibility of holding backup lock into account.