Bug#25813089 - CREATE USER IF NOT EXISTS SHOULD ALWAYS BE                WRITTEN TO THE BINLOGS
DESCRIPTION
===========
MySQL 5.7 introduced a new command CREATE USER IF NOT
EXISTS ... The idea of this command is that it can be run
safely on a master and would not break things irrespective
of whether the user exists on the master or on the slave.
However, the implementation does not write this statement
to the binlog when the user exists on the master which
results in an inconsistency.

ANALYSIS
========
Presently, if the statements CREATE USER IF NOT EXISTS are
a no-op on the master, they are not written to the the
binlog. This results in an inconsistent behavior if a
particular user is present on the slave but not on the
master. Similar issue arises for ALTER USER IF EXISTS
statements.

FIX
===
The fix ensures that the statements having queries like
1. CREATE USER IF NOT EXISTS .. and
2. ALTER USER IF EXISTS ..
are binlogged making sure that replication does not
break and the consistency is maintained across all masters
and slaves.