Bug #18898433: EXTREMELY SLOW PERFORMANCE WITH OUTER JOINS                AND JOIN BUFFER
Issue:
------
The algorithm for join_records is :

1. Generate fully matching records by calling
   join_matching_records(). That is, extensions for which
   the ON condition is true.
2. Call to join_records() recursively to extend these
   matching records for remaining tables.
3. Generate null complemented extensions for those rows in
   the join buffer that do not have matches.
   (Call to join_null_complements())
4. Recursively extend null complemented rows for remaining
   tables. (second call to join_records()).

For joins of the form t1 LEFT JOIN (t2 INNER JOIN t3), that
is the first_inner and last_inner aren't the same, it is
necessary to call join_records() twice to null complement
all the tables.

But for queries of the form t1 LEFT JOIN t2 LEFT JOIN t3,
is the first_inner and last_inner same, the first call to
join_records (Step 2) isn't necessary. Since null
complementing will happen anyway towards the end.

This is problematic because the extra (and recursive) call
to join_records will result in extra reads from the relevant
tables.

Solution:
---------
When there is a single inner table for outer join, skip the
Step 2 (linked join_cache's call to join_records()).

This will avoid both extra processing and I/O.