Bug#25996291 - NEED MORE INFORMATION THAN: PERFORMANCE                SCHEMA DISABLED (REASON: INIT FAILED)
DESCRIPTION
===========
If Performance schema initialization fails, the server log
just reads,
"[Warning] Performance schema disabled (reason: init
failed)."

It doesn't tell the exact reason what caused this failure
making it difficult to overcome the failure.

ANALYSIS
========
initialize_performance_schema() executes several sub-init
functions which are primarily involved in allocating the
memory [with the help of pfs_malloc_array()] for various
buffers. This function just return the status of the inits,
however it would be helpful if something as below is
printed.
"Failed to allocate <x bytes> for buffer <name>"

FIX
===
We can either do this change in all the sub-init functions
or we can do it in the callee function pfs_malloc_array().
Following the 2nd approach and printing the warning now.