Bug#26378589 : CPU LOCKING FEATURE OF THE RECEIVER THREADS IS NOT ENABLED IN LINUX BUILDS
The CPU locking feature is not enabled in the Linux builds due to an
issue in the ndb_configure.cmake's test program used to check the
availability of the sched_setaffinity syscall. This patch fixes that and
one more issue in the ndb_configure.cmake file due to which the
HAVE_FUTEX macro was not being enabled.

Also, due to a bug in the GNU C library's sched_setaffinity call,
requests that change the CPU affinity fails. Due to this bug in the
library, the unset_recv_thread_cpu api call fails in the linux
platform. This bug in the GNU library is fixed from version 2.23. So
this patch makes that version a minimum requirement for the CPU Locking
feature to be enabled.