Bug #26483909: INCORRECT BEHAVIOR FOR QUERY WITH DISTINCT                and ORDER BY ... LIMIT.
Issue:
------
This problem can occur when :
1) Query contains DISTINCT and ORDER BY ... LIMIT.
2) Range access will be used but the conditions are such
that it will still require a temporary table for ORDER BY.


Root cause:
-----------
test_if_skip_sort_order() is called from
JOIN::optimize_distinct_group_order and this returns true
meaning that an index can be used and there is no need for
sorting. But this is called with the flag "no_changes" set
to true, that is no change to the plan will be affected.

When the same function is called from JOIN::test_skip_sort,
the "no_changes" is set to false and hence the plan can be
changed. But since a range access isn't possible for the
relevant index, it returns a false.

This means that m_ordered_index_usage is left with
ORDERED_INDEX_VOID and will cause an assert failure.

Solution:
---------
Ideally JOIN::skip_sort_order should be set to false when
test_if_skip_sort_order() returns a false. This will avoid
optimizing the distinct clause while assuming the ORDER BY
doesn't require a sort.

Other changes:
--------------
It is very confusing that the member variable
'ordered_index_usage' and the values 'ordered_index_*' are
in the lower case. This has been corrected.