Bug #25960230 899 ROWID ALREADY ALLOCATED ERROR UNDER HEAVY  LOAD
Error 899 "rowID already allocated" occurs when a rowID is
available at the primary and unavailable at the backup. An
insert operation seizes the rowID at the primary and fails to
seize it at the backup since it is unavailable.

For multiple operations on the same tuple where the end result
of the transaction is that the tuple is deleted, the last
operation on that tuple is marked as the deallocator of the
rowID. The operation is marked by setting the m_dealloc flag
during COMMIT. When any op COMPLETEs, its m_dealloc flag is
checked, and if set, its rowID is deallocated. In this case,
a transaction T1 contains a READ followed by a DELETE on the
same tuple, while another transaction T2 runs INSERT operations
on the same table.

T1 commits/completes in the unexpected (not incorrect) order:

COMMIT(DELETE)->COMMIT(READ)->COMPLETE(READ)->COMPLETE(DELETE)

- T1: DELETE commits on primary
- T1: READ commits on primary, m_dealloc=1 for READ on primary
- T1: READ completes on primary, deallocates rowID X:Y
- T2: INSERT seizes freed rowID X:Y on primary
- T2: INSERT tries to seize rowID X:Y on backup, gets error 899
  since rowID is not freed yet on backup
- T1: DELETE completes on backup, deallocates rowID X:Y

This is fixed by enforcing a constraint for commits that the
dealloc op must be a write op. Write ops COMPLETE first on the
backup and then on the primary, so the sequence of
          dealloc@backup -> dealloc@primary
is enforced. So with write ops deallocating, a rowID cannot be
freed on the primary before it is freed on the backup.

The fix is that the last write op on the tuple is chosen as the
dealloc op instead of the last read/write op on the tuple. Some
refactoring is also done to separate the 'mark dealloc op' logic
from the locking and commit logic, so that the existing behaviour
for aborts can be retained.